Add elbo loss back into losses file
