+++ alibi_detect/models/tensorflow/tests/test_autoencoder_tf.py
@@ -3,8 +3,7 @@ import pytest
 import tensorflow as tf
 from tensorflow.keras.layers import Dense, InputLayer
 from alibi_detect.models.tensorflow.autoencoder import AE, AEGMM, VAE, VAEGMM, Seq2Seq, EncoderLSTM, DecoderLSTM
-from alibi_detect.models.tensorflow.losses import loss_aegmm
-from alibi_detect.models.tensorflow.prob_losses import loss_vaegmm
+from alibi_detect.models.tensorflow.losses import loss_aegmm, loss_vaegmm
 from alibi_detect.models.tensorflow.trainer import trainer
 
 input_dim = 784
