+++ alibi_detect/models/tensorflow/tests/test_losses_tf.py
@@ -1,7 +1,6 @@
 import numpy as np
 import tensorflow as tf
-from alibi_detect.models.tensorflow.losses import loss_adv_ae, loss_aegmm, loss_distillation
-from alibi_detect.models.tensorflow.prob_losses import elbo, loss_vaegmm
+from alibi_detect.models.tensorflow.losses import loss_adv_ae, loss_aegmm, loss_distillation, elbo, loss_vaegmm
 
 N, K, D, F = 10, 5, 1, 3
 x = np.random.rand(N, F).astype(np.float32)
