+++ alibi_detect/od/vae.py
@@ -5,7 +5,7 @@ import tensorflow_probability as tfp
 from typing import Dict, Tuple
 from alibi_detect.models.tensorflow.autoencoder import VAE
 from alibi_detect.models.tensorflow.trainer import trainer
-from alibi_detect.models.tensorflow.prob_losses import elbo
+from alibi_detect.models.tensorflow.losses import elbo
 from alibi_detect.base import BaseDetector, FitMixin, ThresholdMixin, outlier_prediction_dict
 from alibi_detect.utils.tensorflow.prediction import predict_batch
